Renamed component Map to LocationSearch and added back Map to exclusively render the map
