+++ app/components/modals/RentModal.tsx
@@ -3,7 +3,6 @@
 import axios from "axios";
 import { toast } from "react-hot-toast";
 import { FieldValues, SubmitHandler, useForm } from "react-hook-form";
-import dynamic from "next/dynamic";
 import { useRouter } from "next/navigation";
 import { useMemo, useState } from "react";
 
@@ -17,6 +16,7 @@ import ImageUpload from "../inputs/ImageUpload";
 import Input from "../inputs/Input";
 import Heading from "../Heading";
 import { PreviewImageWithUrl } from "@/app/types";
+import LocationSearch from "../inputs/LocationSearch";
 
 enum STEPS {
   CATEGORY = 0,
@@ -67,14 +67,6 @@ const RentModal = () => {
   const roomCount = watch("roomCount");
   const bathroomCount = watch("bathroomCount");
   const images = watch("images");
-
-  const Map = useMemo(
-    () =>
-      dynamic(() => import("../Map"), {
-        ssr: false,
-      }),
-    [location]
-  );
 
   const setCustomValue = (id: string, value: any) => {
     setValue(id, value, {
@@ -177,7 +169,7 @@ const RentModal = () => {
           subtitle="Help guests find you!"
         />
         <div style={{ height: "calc(35vh + 44px)" }}>
-          <Map
+          <LocationSearch
             value={location}
             onChange={(value) => setCustomValue("location", value)}
             searchString={searchString}
